Handle errors more gracefully
Problem:
Unclear error display when an illegal_url or bearer token is used.

Solution:
URL with trailing slash will immediately raise an exception.
Illegal client_id and secret will generate an auth token that is None
and this will raise an exception on first use.
Logger handling is now fully documented with examples.
Docs generation fixed (spelling plugin was preventing generation of
html)
Archivist unittests split into smaller files as original exceeded
1000 lines.
