+++ examples/create_asset.py
@@ -9,6 +9,7 @@ the credentials, called "arch", then call arch.assets.create() and the asset wil
 from os import getenv
 
 from archivist.archivist import Archivist
+from archivist.logger import set_logger
 from archivist.proof_mechanism import ProofMechanism
 from archivist.utils import get_auth
 
@@ -74,6 +75,10 @@ def main():
     create an example archivist connection and create an asset.
 
     """
+    # optional call to set the logger level for all subsystems. The argumant can
+    # be either "INFO" or "DEBUG". For more sophisticated logging control see the
+    # documentation.
+    set_logger("INFO")
 
     # client id and client secret is obtained from the appidp endpoint - see the
     # application registrations example code in examples/applications_registration.py
