+++ unittests/testarchivist.py
@@ -9,11 +9,9 @@ from unittest import TestCase, mock
 from archivist.archivist import Archivist
 from archivist.constants import ROOT, HEADERS_TOTAL_COUNT, HEADERS_RETRY_AFTER
 from archivist.errors import (
-    ArchivistBadFieldError,
     ArchivistBadRequestError,
-    ArchivistDuplicateError,
+    ArchivistError,
     ArchivistHeaderError,
-    ArchivistNotFoundError,
     ArchivistTooManyRequestsError,
 )
 from archivist.logger import set_logger
@@ -43,12 +41,24 @@ RESPONSE = {
     "expires_in": 660,
     "token_type": "Bearer",
 }
+NONE_RESPONSE = {
+    "access_token": None,
+    "expires_in": 660,
+    "token_type": "Bearer",
+}
 
 
 class TestArchivist(TestCase):
     """
     Test Archivist class
     """
+
+    def test_archivist_illegal_urlk(self):
+        """
+        Test illegal url
+        """
+        with self.assertRaises(ArchivistError):
+            arch = Archivist("url/", "authauthauth")
 
     def test_archivist(self):
         """
@@ -150,6 +160,16 @@ class TestArchivist(TestCase):
                 ACCESS_TOKEN,
                 msg="Incorrect auth",
             )
+
+    def test_archivist_none_token(self):
+        """
+        Test archivist creation with none token
+        """
+        arch = Archivist("url", (CLIENT_ID, CLIENT_SECRET))
+        with mock.patch.object(arch.appidp, "token") as mock_token:
+            mock_token.return_value = NONE_RESPONSE
+            with self.assertRaises(ArchivistError):
+                _ = arch.auth
 
     def test_archivist_copy(self):
         """
@@ -467,516 +487,3 @@ class TestArchivistCount(TestArchivistMethods):
                 1,
                 msg="incorrect count",
             )
-
-
-class TestArchivistList(TestArchivistMethods):
-    """
-    Test Archivist list method
-    """
-
-    def test_list(self):
-        """
-        Test default list method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            responses = list(self.arch.list("path/path", "things"))
-            self.assertEqual(
-                len(responses),
-                1,
-                msg="incorrect number of responses",
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": None,
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-    def test_list_with_error(self):
-        """
-        Test default list method with error
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                400,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            with self.assertRaises(ArchivistBadRequestError):
-                responses = list(self.arch.list("path/path", "things"))
-
-    def test_list_with_bad_field(self):
-        """
-        Test default list method with error
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            with self.assertRaises(ArchivistBadFieldError):
-                responses = list(self.arch.list("path/path", "badthings"))
-
-    def test_list_with_headers(self):
-        """
-        Test default list method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            responses = list(
-                self.arch.list(
-                    "path/path",
-                    "things",
-                    headers={"headerfield1": "headervalue1"},
-                )
-            )
-            self.assertEqual(
-                len(responses),
-                1,
-                msg="incorrect number of responses",
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                                "headerfield1": "headervalue1",
-                            },
-                            "params": None,
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-    def test_list_with_params(self):
-        """
-        Test default list method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            responses = list(
-                self.arch.list(
-                    "path/path",
-                    "things",
-                    params={"paramsfield1": "paramsvalue1"},
-                )
-            )
-            self.assertEqual(
-                len(responses),
-                1,
-                msg="incorrect number of responses",
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "verify": True,
-                            "params": {"paramsfield1": "paramsvalue1"},
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-    def test_list_with_page_size(self):
-        """
-        Test default list method
-        """
-        values = ("value10", "value11")
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": values[0],
-                    },
-                    {
-                        "field1": values[1],
-                    },
-                ],
-            )
-            responses = list(
-                self.arch.list(
-                    "path/path",
-                    "things",
-                    page_size=2,
-                )
-            )
-            self.assertEqual(
-                len(responses),
-                2,
-                msg="incorrect number of responses",
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": {"page_size": 2},
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-            for i, r in enumerate(responses):
-                self.assertEqual(
-                    r["field1"],
-                    values[i],
-                    msg="Incorrect response body value",
-                )
-
-    def test_list_with_multiple_pages(self):
-        """
-        Test default list method
-        """
-        values = ("value10", "value11", "value12", "value13")
-        paging = ({"page_size": 2}, {"page_size": 2, "page_token": "token"})
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.side_effect = [
-                MockResponse(
-                    200,
-                    things=[
-                        {
-                            "field1": values[0],
-                        },
-                        {
-                            "field1": values[1],
-                        },
-                    ],
-                    next_page_token="token",
-                ),
-                MockResponse(
-                    200,
-                    things=[
-                        {
-                            "field1": values[2],
-                        },
-                        {
-                            "field1": values[3],
-                        },
-                    ],
-                ),
-            ]
-            responses = list(
-                self.arch.list(
-                    "path/path",
-                    "things",
-                    page_size=2,
-                )
-            )
-            self.assertEqual(
-                len(responses),
-                4,
-                msg="incorrect number of responses",
-            )
-            for i, a in enumerate(mock_get.call_args_list):
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": paging[i],
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-            for i, r in enumerate(responses):
-                self.assertEqual(
-                    r["field1"],
-                    values[i],
-                    msg="Incorrect response body value",
-                )
-
-    def test_list_with_multiple_pages_and_params(self):
-        """
-        Test default list method
-        """
-        params = {"field2": "value2"}
-        values = ("value10", "value11", "value12", "value13")
-        paging = (
-            {**params, "page_size": 2},
-            {"page_size": 2, "page_token": "token"},
-        )
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.side_effect = [
-                MockResponse(
-                    200,
-                    things=[
-                        {
-                            "field1": values[0],
-                        },
-                        {
-                            "field1": values[1],
-                        },
-                    ],
-                    next_page_token="token",
-                ),
-                MockResponse(
-                    200,
-                    things=[
-                        {
-                            "field1": values[2],
-                        },
-                        {
-                            "field1": values[3],
-                        },
-                    ],
-                ),
-            ]
-            responses = list(
-                self.arch.list(
-                    "path/path",
-                    "things",
-                    page_size=2,
-                    params=params,
-                )
-            )
-            self.assertEqual(
-                len(responses),
-                4,
-                msg="incorrect number of responses",
-            )
-            for i, a in enumerate(mock_get.call_args_list):
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": paging[i],
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-            for i, r in enumerate(responses):
-                self.assertEqual(
-                    r["field1"],
-                    values[i],
-                    msg="Incorrect response body value",
-                )
-
-    def test_list_with_429(self):
-        """
-        Test list method with error
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(429)
-            with self.assertRaises(ArchivistTooManyRequestsError):
-                things = list(self.arch.list("path/path", "things"))
-
-    def test_list_with_429_retry_and_fail(self):
-        """
-        Test list method with 429 retry and fail
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.side_effect = (
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(429),
-            )
-            with self.assertRaises(ArchivistTooManyRequestsError):
-                things = list(self.arch.list("path/path", "things"))
-
-    def test_list_with_429_retry_and_retries_fail(self):
-        """
-        Test list method with 429 retry and retries_fail
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.side_effect = (
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-            )
-            with self.assertRaises(ArchivistTooManyRequestsError):
-                things = list(self.arch.list("path/path", "things"))
-
-    def test_list_with_429_retry_and_success(self):
-        """
-        Test list method with 429 retry and success
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.side_effect = (
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(429, headers={HEADERS_RETRY_AFTER: 0.1}),
-                MockResponse(
-                    200,
-                    headers={HEADERS_TOTAL_COUNT: 1},
-                    things=[
-                        {
-                            "field1": "value1",
-                        },
-                    ],
-                ),
-            )
-            things = list(self.arch.list("path/path", "things"))
-            self.assertEqual(
-                len(things),
-                1,
-                msg="incorrect number of things",
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": None,
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-
-class TestArchivistSignature(TestArchivistMethods):
-    """
-    Test Archivist get_by_signature method
-    """
-
-    def test_get_by_signature(self):
-        """
-        Test default get_by_signature method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            entity = self.arch.get_by_signature(
-                "path/path", "things", {"field1": "value1"}
-            )
-            for a in mock_get.call_args_list:
-                self.assertEqual(
-                    tuple(a),
-                    (
-                        (f"url/{ROOT}/path/path",),
-                        {
-                            "headers": {
-                                "authorization": "Bearer authauthauth",
-                            },
-                            "params": {"field1": "value1", "page_size": 2},
-                            "verify": True,
-                        },
-                    ),
-                    msg="GET method called incorrectly",
-                )
-
-    def test_get_by_signature_not_found(self):
-        """
-        Test default get_by_signature method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[],
-            )
-            with self.assertRaises(ArchivistNotFoundError):
-                entity = self.arch.get_by_signature(
-                    "path/path", "things", {"field1": "value1"}
-                )
-
-    def test_get_by_signature_duplicate(self):
-        """
-        Test default get_by_signature method
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            with self.assertRaises(ArchivistDuplicateError):
-                entity = self.arch.get_by_signature(
-                    "path/path", "things", {"field1": "value1"}
-                )
-
-    def test_get_by_signature_with_bad_field(self):
-        """
-        Test default list method with error
-        """
-        with mock.patch.object(self.arch._session, "get") as mock_get:
-            mock_get.return_value = MockResponse(
-                200,
-                things=[
-                    {
-                        "field1": "value1",
-                    },
-                ],
-            )
-            with self.assertRaises(ArchivistBadFieldError):
-                entity = self.arch.get_by_signature(
-                    "path/path", "badthings", {"field1": "value1"}
-                )
